+++ WEB-INF/src/gov/alaska/dggs/igneous/model/Unit.java
@@ -7,10 +7,10 @@ public class Unit implements Serializable
 {
 	private static final long serialVersionUID = 1L;
 
-	private long id;
+	private int id;
 	private String name, abbr, description;
 
-	public long getID(){ return id; }
+	public int getID(){ return id; }
 	public String getName(){ return name; }
 	public String getAbbr(){ return abbr; }
 	public String getDescription(){ return description; }
